a lot of clean + documentation
